pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "wheel"
  ]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-unicore-integration"
version = "0.1.0"
authors = [
  { name="Christian Böttcher", email="c.boettcher@fz-juelich.de" },
]
description = "Running Unicore Jobs from airflow DAGs."
readme = "README.rst"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Apache Airflow :: Provider",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]



dependencies = [
    "pyunicore>=1.0.0",
    "apache-airflow>=2.10.0"
]

[project.urls]
Homepage = "https://github.com/UNICORE-EU/airflow-unicore-integration"
Issues = "https://github.com/UNICORE-EU/airflow-unicore-integration/issues"

[project.entry-points."apache_airflow_provider"]
provider_info = "airflow_unicore_integration:get_provider_info"

[project.entry-points.'airflow.policy']
_ = 'airflow_unicore_integration.policies'

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.black]
line-length = 100

[tool.flake8]
max-line-length = 100
ignore = [
 "N999",
 "E501",
 "W503"
]
per-file-ignores = [
    "__init__.py:F401,E501",
    "_version.py:E203",
]

[tool.mypy]
ignore_missing_imports=true
